Rerun make for StaticLib

+++ OpenRatesSiriStaticLib/bo/StructsAndEnums.cs
@@ -1,11 +1,10 @@
-using System;
 using ObjCRuntime;
 
 namespace OpenRates.iOS.Siri
 {
 	[Watch (5,0), NoTV, NoMac, iOS (12,0)]
 	[Native]
-	public enum Currency : nint
+	public enum Currency : long
 	{
 		Unknown = 0,
 		Eur = 1,
@@ -20,7 +19,7 @@ namespace OpenRates.iOS.Siri
 
 	[Watch (5,0), NoTV, NoMac, iOS (12,0)]
 	[Native]
-	public enum ConvertAmountToCurrencyIntentResponseCode : nint
+	public enum ConvertAmountToCurrencyIntentResponseCode : long
 	{
 		Unspecified = 0,
 		Ready,
@@ -33,7 +32,7 @@ namespace OpenRates.iOS.Siri
 
 	[Watch (6,0), iOS (13,0)]
 	[Native]
-	public enum ConvertAmountToCurrencyAmountUnsupportedReason : nint
+	public enum ConvertAmountToCurrencyAmountUnsupportedReason : long
 	{
 		NegativeNumbersNotSupported = 1,
 		GreaterThanMaximumValue,
@@ -42,14 +41,14 @@ namespace OpenRates.iOS.Siri
 
 	[Watch (6,0), iOS (13,0)]
 	[Native]
-	public enum ConvertAmountToCurrencyTargetCurrencyUnsupportedReason : nint
+	public enum ConvertAmountToCurrencyTargetCurrencyUnsupportedReason : long
 	{
 		ConvertAmountToCurrencyTargetCurrencyUnsupportedReasonUnknownCurrency = 1
 	}
 
 	[Watch (5,0), NoTV, NoMac, iOS (12,0)]
 	[Native]
-	public enum GetConversionRateIntentResponseCode : nint
+	public enum GetConversionRateIntentResponseCode : long
 	{
 		Unspecified = 0,
 		Ready,
@@ -62,14 +61,14 @@ namespace OpenRates.iOS.Siri
 
 	[Watch (6,0), iOS (13,0)]
 	[Native]
-	public enum GetConversionRateBaseCurrencyUnsupportedReason : nint
+	public enum GetConversionRateBaseCurrencyUnsupportedReason : long
 	{
 		GetConversionRateBaseCurrencyUnsupportedReasonUnknownCurrency = 1
 	}
 
 	[Watch (6,0), iOS (13,0)]
 	[Native]
-	public enum GetConversionRateTargetCurrencyUnsupportedReason : nint
+	public enum GetConversionRateTargetCurrencyUnsupportedReason : long
 	{
 		GetConversionRateTargetCurrencyUnsupportedReasonUnknowCurrency = 1
 	}
